Add windows check
